Implemented a initial HP for SWActor(s)
This will be used when healing players and NPCs. If a player attempts to heal a NPC (Droid) that has full HP, the plan is to output a full HP check.
Also, healing is planned to completely heal a NPC, so for the oil can with a capacity of 100, it can heal a Droid 2x over (as current)