Migrate wield to be an Action
at the moment it doesnt print the Wield affordance when i pick up a weapon - it's just empty and tells me to move,. trying to bug fix now.
logic is in place to allow wielding of a lightsaber, or other force weapons depending on the actors forcePower and the WIELD FORCE PWR REQ which is set to 20 by default.