Leia added to Lukes' follow list upon reach, Leia follows Luke around the Death Star
Implemented the following of Leia when the player reaches her. No action necessary = as she will automatically check if Luke is nearby - and if he is - always follow him.
Leia is also added to the follower list of Luke - so along with R2D2, constitutes a win (to be implemented)