Potential implementation of travel working (WIP)
Potential working transfer of Luke to YAvin 4 from Tatooine implemented. Definitely needs work, but the esssence of world transfer logic is implemented.
Can see that when trying to go back to Tatooine, the world clones itself which is due to the current way the function is implemented at the moment, will work on it!